requirements.txt
django
django-ckeditor
sorl-thumbnail
pillow
flake8
djangorestframework
django-cors-headers